Refactor streaming system into usePatchLoader
This at least organizes/hides some of the craziness
